requirements.txt
customtkinter
yfinance
pandas
matplotlib
google-generativeai
python-dotenv
pydantic